Refactor configurator
#83
Change config order format, simplify changing config functions, combine scattered configurator methods into one file
