+++ device/config/configurator.go
@@ -1,47 +1,51 @@
 package config
 
 import (
+	"fmt"
+	"strings"
+
+	log "github.com/Sirupsen/logrus"
+	"github.com/pkg/errors"
+	"github.com/xshellinc/tools/constants"
 	"github.com/xshellinc/tools/dialogs"
+	"github.com/xshellinc/tools/lib/help"
+	"github.com/xshellinc/tools/lib/ping"
 	"github.com/xshellinc/tools/lib/ssh_helper"
+	"sort"
 )
 
 // Default Configurator constants that are describe a specific configuration option
 const (
-	Locale = iota
-	Keymap
-	Wifi
-	Interface
-	DNS
+	Locale    = "Locale"
+	Keymap    = "Keymap"
+	Wifi      = "Wifi"
+	Interface = "Interface"
+	DNS       = "DNS"
+	SSH       = "SSH"
 )
-
-// consts are a const strings representation
-var consts = [...]string{
-	"Locale",
-	"Keymap",
-	"Wifi",
-	"Interface",
-	"DNS",
-}
-
-// GetConstLiteral gets a literal from configurator.const
-func GetConstLiteral(v int) string {
-	return consts[v]
-}
 
 type (
 	// configurator is a container of a mutual storage and order of CallbackFn
 	configurator struct {
 		storage map[string]interface{}
-		order   []*CallbackFn
+		order   map[string]*callbackFn
 	}
 
-	// Function is a function with an input parameter of configurator's `storage`
-	Function func(map[string]interface{}) error
+	// cb is a function with an input parameter of configurator's `storage`
+	cb func(map[string]interface{}) error
 
 	// CallbackFn is an entity with Config and Apply function
-	CallbackFn struct {
-		Config Function
-		Apply  Function
+	callbackFn struct {
+		Config cb
+		Apply  cb
+	}
+
+	// Interfaces represents network interfaces used to setup devices
+	Interfaces struct {
+		Address string
+		Gateway string
+		Netmask string
+		DNS     string
 	}
 )
 
@@ -50,7 +54,7 @@ func New(ssh ssh_helper.Util) *configurator {
 	storage := make(map[string]interface{})
 
 	// default
-	order := make([]*CallbackFn, 0)
+	order := make(map[string]*callbackFn)
 
 	storage["ssh"] = ssh
 
@@ -61,29 +65,32 @@ func New(ssh ssh_helper.Util) *configurator {
 func NewDefault(ssh ssh_helper.Util) *configurator {
 	config := New(ssh)
 	// add default callbacks
-	config.order = append(config.order, NewCallbackFn(SetLocale, SaveLocale))
-	config.order = append(config.order, NewCallbackFn(SetKeyboard, SaveKeyboard))
-	config.order = append(config.order, NewCallbackFn(SetWifi, SaveWifi))
-	config.order = append(config.order, NewCallbackFn(SetInterface, SaveInterface))
-	config.order = append(config.order, NewCallbackFn(SetSecondaryDNS, SaveSecondaryDNS))
+	config.order[Locale] = NewCallbackFn(SetLocale, SaveLocale)
+	config.order[Keymap] = NewCallbackFn(SetKeyboard, SaveKeyboard)
+	config.order[Wifi] = NewCallbackFn(SetWifi, SaveWifi)
+	config.order[Interface] = NewCallbackFn(SetInterface, SaveInterface)
+	config.order[DNS] = NewCallbackFn(SetSecondaryDNS, SaveSecondaryDNS)
 	return config
 }
 
 // NewCallbackFn creates a new CallbackFn with 2 Function parameters
-func NewCallbackFn(config Function, apply Function) *CallbackFn {
-	return &CallbackFn{config, apply}
+func NewCallbackFn(config cb, apply cb) *callbackFn {
+	return &callbackFn{config, apply}
 }
 
 // Setup triggers all CallbackFn Config functions
 func (c *configurator) Setup() error {
-	if dialogs.YesNoDialog("Would you like to configure your board?") {
-		for _, o := range c.order {
-			if (*o).Config == nil {
-				continue
-			}
-			if err := o.Config(c.storage); err != nil {
-				return err
-			}
+	var keys []string
+	for k := range c.order {
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+	for _, k := range keys {
+		if (*c.order[k]).Config == nil {
+			continue
+		}
+		if err := c.order[k].Config(c.storage); err != nil {
+			return err
 		}
 	}
 
@@ -106,32 +113,296 @@ func (c *configurator) Write() error {
 }
 
 // AddConfigFn
-func (c *configurator) AddConfigFn(ccf *CallbackFn) {
-	c.order = append(c.order, ccf)
+func (c *configurator) AddConfigFn(name string, ccf *callbackFn) {
+	c.order[name] = ccf
 }
 
-// SetConfigFn sets CallbackFn of a specified number of the array
-func (c *configurator) SetConfigFn(num int, ccf *CallbackFn) {
-	c.order[num] = ccf
+// SetConfigFn sets CallbackFn of the specified name
+func (c *configurator) SetConfigFn(name string, ccf *callbackFn) {
+	c.order[name] = ccf
 }
 
-// GetConfigFn returns GetConfigFn from the array by a number
-func (c *configurator) GetConfigFn(num int) *CallbackFn {
-	return c.order[num]
+// GetConfigFn returns configuration function by it's name
+func (c *configurator) GetConfigFn(name string) *callbackFn {
+	return c.order[name]
 }
 
 // RemoveConfigFn removes CallbackFn from order
-func (c *configurator) RemoveConfigFn(num int) {
-	if num == len(c.order)-1 {
-		c.order = c.order[:num]
-	}
-	if num == 0 {
-		c.order = c.order[1:]
-	}
-
-	c.order = append(c.order[:num], c.order[num+1:]...)
+func (c *configurator) RemoveConfigFn(name string) {
+	delete(c.order, name)
 }
 
 func (c *configurator) StoreValue(name string, value interface{}) {
 	c.storage[name] = value
+}
+
+// SetLocale is a default method to with dialog to configure the locale
+func SetLocale(storage map[string]interface{}) error {
+
+	fmt.Println("[+] Default language: ", constants.DefaultLocale)
+	if dialogs.YesNoDialog("Change default language?") {
+		inp := dialogs.GetSingleAnswer("New locale: ", dialogs.CreateValidatorFn(constants.ValidateLocale))
+
+		arr := constants.GetLocale(inp)
+
+		var l string
+		if len(arr) == 1 {
+			l = arr[0].Locale
+		} else {
+			l = arr[dialogs.SelectOneDialog("Please select a locale from a list: ", arr.Strings())].Locale
+		}
+
+		storage[Locale] = l
+	}
+
+	return nil
+}
+
+// SaveLocale is a default method to save locale into the image
+func SaveLocale(storage map[string]interface{}) error {
+
+	if _, ok := storage[Locale]; !ok {
+		return nil
+	}
+
+	ssh, ok := storage["ssh"].(ssh_helper.Util)
+	if !ok {
+		return errors.New("Cannot get ssh config")
+	}
+
+	fp := help.AddPathSuffix("unix", constants.MountDir, constants.ISAAX_CONF_DIR, constants.LocaleF)
+	data := fmt.Sprintf(constants.Language+constants.LocaleLang, storage[Locale], storage[Locale])
+
+	_, eut, err := ssh.Run(fmt.Sprintf(`echo "%s" > %s`, data, fp))
+	if err != nil {
+		return errors.New(err.Error() + ":" + eut)
+	}
+
+	fp = help.AddPathSuffix("unix", constants.MountDir, constants.ISAAX_CONF_DIR, "environment")
+	data = fmt.Sprintf(constants.LocaleAll, storage[Locale])
+
+	_, eut, err = ssh.Run(fmt.Sprintf(`echo "%s" > %s`, data, fp))
+	if err != nil {
+		return errors.New(err.Error() + ":" + eut)
+	}
+
+	return nil
+}
+
+// SetKeyboard is a default method to with dialog to configure the keymap
+func SetKeyboard(storage map[string]interface{}) error {
+	var (
+		locale string
+		ok     bool
+	)
+
+	if locale, ok = storage[Locale].(string); ok {
+		if i := strings.IndexAny(locale, "_."); i >= 0 {
+			locale = locale[:i]
+		}
+	}
+
+	fmt.Println("[+] Default keyboard layout: ", constants.DefaultKeymap)
+
+	if dialogs.YesNoDialog("Change default keyboard layout?") {
+		inp := dialogs.GetSingleAnswer("New keyboard layout: ",
+			dialogs.CreateValidatorFn(func(layout string) error { return constants.ValidateLayout(locale, layout) }))
+
+		arr := constants.GetLayout(locale, inp)
+
+		var l string
+		if len(arr) == 1 {
+			l = arr[0].Layout
+		} else {
+			l = arr[dialogs.SelectOneDialog("Please select a layout from a list: ", arr.Strings())].Layout
+		}
+
+		storage[Keymap] = fmt.Sprintf(constants.KeyMap, l)
+	}
+
+	return nil
+}
+
+// SaveKeyboard is a default method to save KEYMAP into the image
+func SaveKeyboard(storage map[string]interface{}) error {
+
+	if _, ok := storage[Keymap]; !ok {
+		return nil
+	}
+
+	ssh, ok := storage["ssh"].(ssh_helper.Util)
+	if !ok {
+		return errors.New("Cannot get ssh config")
+	}
+
+	fp := help.AddPathSuffix("unix", constants.MountDir, constants.ISAAX_CONF_DIR, constants.KeyboardF)
+	data := storage[Keymap]
+
+	_, eut, err := ssh.Run(fmt.Sprintf(`echo "%s" > %s`, data, fp))
+	if err != nil {
+		return errors.New(err.Error() + ":" + eut)
+	}
+
+	return nil
+}
+
+// SetWifi is a dialog asking to configure wpa supplicant
+func SetWifi(storage map[string]interface{}) error {
+	if dialogs.YesNoDialog("Would you like to configure your Wi-Fi?") {
+		storage[Wifi+"_name"] = dialogs.GetSingleAnswer("WIFI SSID name: ", dialogs.EmptyStringValidator)
+		storage[Wifi+"_pass"] = []byte(dialogs.WiFiPassword())
+	}
+
+	return nil
+}
+
+// SaveWifi is a default method to save wpa_supplicant for the wifi connection
+func SaveWifi(storage map[string]interface{}) error {
+
+	if _, ok := storage[Wifi+"_name"]; !ok {
+		return nil
+	}
+
+	ssh, ok := storage["ssh"].(ssh_helper.Util)
+	if !ok {
+		return errors.New("Cannot get ssh config")
+	}
+
+	fp := help.AddPathSuffix("unix", constants.MountDir, constants.ISAAX_CONF_DIR, "wpa_supplicant", constants.WPAsupplicant)
+	data := fmt.Sprintf(constants.WPAconf, storage[Wifi+"_name"], storage[Wifi+"_pass"])
+
+	_, eut, err := ssh.Run(fmt.Sprintf(`echo "%s" > %s`, data, fp))
+	if err != nil {
+		return errors.New(err.Error() + ":" + eut)
+	}
+
+	return nil
+}
+
+// SetInterface is a dialog asking to setup user Interfaces for the static ip functionality
+func SetInterface(storage map[string]interface{}) error {
+	log.WithField("type", "default").Debug("setInterface")
+	device := []string{"eth0", "wlan0"}
+	i := Interfaces{
+		Address: "192.168.0.254",
+		Netmask: "255.255.255.0",
+		Gateway: "192.168.0.1",
+		DNS:     "192.168.0.1",
+	}
+
+	if dialogs.YesNoDialog("Would you like to assign static IP address for your device?") {
+		fmt.Println("[+] Available network interface: ")
+		num := dialogs.SelectOneDialog("Please select a network interface: ", device)
+		fmt.Println("[+] ********NOTE: ADJUST THESE VALUES ACCORDING TO YOUR LOCAL NETWORK CONFIGURATION********")
+
+		fmt.Printf("[+] Current values are:\n \t[+] Address:%s\n\t[+] Gateway:%s\n\t[+] Netmask:%s\n\t[+] DNS:%s\n",
+			i.Address, i.Gateway, i.Netmask, i.DNS)
+
+		if dialogs.YesNoDialog("Change values?") {
+			AskInterfaceParams(&i)
+		}
+
+		switch device[num] {
+		case "eth0":
+			storage[Interface] = fmt.Sprintf(constants.InterfaceETH, i.Address, i.Netmask, i.Gateway, i.DNS)
+			fmt.Println("[+]  Ethernet interface configuration was updated")
+		case "wlan0":
+			storage[Interface] = fmt.Sprintf(constants.InterfaceWLAN, i.Address, i.Netmask, i.Gateway, i.DNS)
+			fmt.Println("[+]  wifi interface configuration was updated")
+		}
+
+	}
+
+	return nil
+}
+
+// SaveInterface is a default method to save user Interfaces into the image
+func SaveInterface(storage map[string]interface{}) error {
+
+	if _, ok := storage[Interface]; !ok {
+		return nil
+	}
+
+	ssh, ok := storage["ssh"].(ssh_helper.Util)
+	if !ok {
+		return errors.New("Cannot get ssh config")
+	}
+
+	fp := help.AddPathSuffix("unix", constants.MountDir, constants.ISAAX_CONF_DIR, "network", constants.InterfacesF)
+
+	_, eut, err := ssh.Run(fmt.Sprintf(`echo "%s" > %s`, storage[Interface], fp))
+	if err != nil {
+		return errors.New(err.Error() + ":" + eut)
+	}
+
+	return nil
+}
+
+// SetSecondaryDNS is a dialog asking to set 8.8.8.8 DNS
+func SetSecondaryDNS(storage map[string]interface{}) error {
+	if dialogs.YesNoDialog("Add Google DNS as a secondary NameServer") {
+		storage[DNS] = true
+		return nil
+	}
+
+	return nil
+}
+
+// SaveSecondaryDNS is a default method to set 8.8.8.8 as a secondary DNS
+func SaveSecondaryDNS(storage map[string]interface{}) error {
+
+	if _, ok := storage[DNS]; !ok {
+		return nil
+	}
+
+	ssh, ok := storage["ssh"].(ssh_helper.Util)
+	if !ok {
+		return errors.New("Cannot get ssh config")
+	}
+
+	fp := help.AddPathSuffix("unix", constants.MountDir, constants.ISAAX_CONF_DIR, "dhcp", "dhclient.conf")
+	command := "append domain-name-servers 8.8.8.8, 8.8.4.4;"
+
+	_, eut, err := ssh.Run(fmt.Sprintf(`echo "%s" >> %s`, command, fp))
+	if err != nil {
+		return errors.New(err.Error() + ":" + eut)
+	}
+
+	return nil
+}
+
+// SetInterfaces is a set of dialog to set user `Interfaces`
+func AskInterfaceParams(i *Interfaces) {
+	loop := true
+	retries := 5
+
+	var ip string
+
+	for retries > 0 && loop {
+		job := help.NewBackgroundJob()
+		ip = dialogs.GetSingleAnswer("IP address of the device: ", dialogs.IpAddressValidator)
+
+		go func() {
+			defer job.Close()
+			loop = !ping.PingIp(ip) //returns false on ping success
+			if loop {
+				fmt.Printf("\n[-] Sorry, device with %s already exists on the network", ip)
+			}
+
+			retries--
+		}()
+		help.WaitJobAndSpin("Checking availability", job)
+	}
+
+	if retries == 0 {
+		if dialogs.YesNoDialog("Do you want to try again?") {
+			AskInterfaceParams(i)
+			return
+		}
+	}
+	i.Address = ip
+	fmt.Println("[+] Using IP:", ip)
+	i.Gateway = dialogs.GetSingleAnswer("Please enter your gateway: ", dialogs.IpAddressValidator)
+	i.Netmask = dialogs.GetSingleAnswer("Please enter your netmask: ", dialogs.IpAddressValidator)
+	i.DNS = dialogs.GetSingleAnswer("Please enter your dns server: ", dialogs.IpAddressValidator)
 }
